Moved logic to extract vc to util class.
